+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+- fix callbacks context was nil
+- fix `retries[:enabled]` flag was no working when was `false`
+
 ## [0.1.0] - 2024-08-30
 
 - Add possibility to verify webhook OFFLINE